requirements.txt
Flask==3.1.2
Jinja2==3.1.6
MarkupSafe==3.0.3
PyYAML==6.0.2
Pygments==2.19.2
SQLAlchemy==2.0.34
Werkzeug==3.1.5
annotated-types==0.7.0
autoflake==2.3.1
black==26.1.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
duckdb==1.1.0
google-ads==29.0.0
google-api-core==2.29.0
google-auth-oauthlib==1.2.4
google-auth==2.48.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio-status==1.78.0
grpcio==1.78.0
idna==3.11
itsdangerous==2.2.0
markdown-it-py==4.0.0
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==2.4.2
oauthlib==3.3.1
packaging==26.0
pandas==2.2.2
pathspec==1.0.4
platformdirs==4.9.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytokens==0.4.1
pytz==2025.2
requests-oauthlib==2.0.0
requests==2.32.5
rich==13.9.2
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
typer==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
